add authorization

+++ app/product/controller.js
@@ -4,10 +4,19 @@ const Product = require("./model");
 const Category = require("../category/model");
 const config = require("../config");
 const Tag = require("../tag/model");
+const { policyFor } = require("../policy");
 
 async function store(req, res, next) {
   try {
+    let policy = policyFor(req.user);
     let payload = req.body;
+
+    if (!policy.can("create", "Product")) {
+      return res.json({
+        error: 1,
+        message: `Anda tidak memiliki akses untuk membuat produk`,
+      });
+    }
 
     if (payload.category) {
       let category = await Category.findOne({ name: { $regex: payload.category, $options: "i" } });
@@ -82,6 +91,8 @@ async function index(req, res, next) {
       };
     }
 
+    let count = await Product.find(criteria).countDocuments();
+
     if (category.length) {
       category = await Category.findOne({ name: { $regex: `${category}` }, $options: "i" });
 
@@ -96,7 +107,7 @@ async function index(req, res, next) {
     }
 
     let products = await Product.find(criteria).limit(parseInt(limit)).skip(parseInt(skip)).populate("category").populate("tags");
-    return res.json(products);
+    return res.json({ data: products, count });
   } catch (err) {
     next(err);
   }
@@ -104,7 +115,15 @@ async function index(req, res, next) {
 
 async function update(req, res, next) {
   try {
+    let policy = policyFor(req.user);
     let payload = req.body;
+
+    if (!policy.can("update", "Product")) {
+      return res.json({
+        error: 1,
+        message: `Anda tidak memiliki akses untuk mengupdate produk`,
+      });
+    }
 
     if (payload.category) {
       let category = await Category.findOne({ name: { $regex: payload.category, $options: "i" } });
@@ -176,6 +195,13 @@ async function update(req, res, next) {
 
 async function destroy(req, res, next) {
   try {
+    let policy = policyFor(req.user);
+    if (!policy.can("delete", "Product")) {
+      return res.json({
+        error: 1,
+        message: "Anda tidak memiliki akses untuk menghapus produk",
+      });
+    }
     let product = await Product.findOneAndDelete({ _id: req.params.id });
     let currentImage = `${config.rootPath}/public/upload${product.image_url}`;
     if (fs.existsSync(currentImage)) {
